requirements.txt
fastapi
==0.109.2
uvicorn[standard]==0.27.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
bcrypt==4.1.2
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1
sqlalchemy==2.0.25
pymysql==1.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0

aiofiles==23.2.1